% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{abundance}
\alias{abundance}
\title{Abundance}
\usage{
abundance(taxa.wide)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}
}
\value{
A count of the total number of organisms observed.
}
\description{
Abundance
}
