% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_metrics.R
\name{all_diversity}
\alias{all_diversity}
\title{Run All Diversity Metrics}
\usage{
all_diversity(master, long.df, class.wide, order.wide, family.wide, tribe.wide,
  genus.wide, taxa.rank.df, taxa.rank)
}
\arguments{
\item{master}{= The master taxa list containing all of the appropriate
taxonomic resolutions and the appropriate taxa attributes. A master taxa list
is contained with the BIBI package.}

\item{long.df}{= Taxonomic data in long format}

\item{class.wide}{= Class level wide taxonomic data frame}

\item{order.wide}{= Order level wide taxonomic data frame}

\item{family.wide}{= Family level wide taxonomic data frame}

\item{taxa.rank.df}{= Specified taxonomic level wide taxonomic data frame}

\item{taxa.rank}{= Taxonomic level ("FAMILY" or "GENUS") (Default = "FAMILY")}
}
\value{
All diversity metric values
}
\description{
Run All Diversity Metrics
}
