% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_metrics.R
\name{all_metrics}
\alias{all_metrics}
\title{Run All Metrics}
\usage{
all_metrics(long.df, master.df, taxa.rank)
}
\arguments{
\item{long.df}{= Taxonomic counts in a long data format.}

\item{master.df}{= The master taxa list contains taxonomic ranks from Phylum
to species and known taxonomic attributes.}

\item{taxa.rank}{= The lowest taxonomic rank ("ORDER", "FAMILY", or "GENUS")
used to calculate the metrics.  If the majority of your taxa are identified
to the family level, then it would be inappropriate to perform metric
calculations at the genus level.}
}
\value{
Calculates all of applicable and available metrics in the package.
}
\description{
Run All Metrics
}
