% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance_metrics.R
\name{becks}
\alias{becks}
\title{Beck's Index}
\usage{
becks(taxa.wide, taxa.rank, master.df, beck.version = 1)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}

\item{taxa.rank}{= The taxonomic rank used to perform the analysis. You must
sepecify either "FAMILY" or "GENUS" (Defualt = "FAMILY").}

\item{master.df}{= A master taxa list including taxonomic ranks Phylum through
the specified taxonomic rank (Family or Genus) and the an
associated list of tolerance values. The default is set to the master taxa
list included in the BIBI package.  The master taxa list can be viewed with
the following script: master.df <- data(master)}

\item{beck.version}{= The version of Beck's Index specified as 1 or 3.  The
defualt is 1, the orginal Beck's Index:
2(S1) + 1(S2) + 3(S3)
where:
S# = Taxa richness at each level of Beck's classifications (0-3).
Beck's Index version 3 is calculated
as specified by PADNR:
3(S0) + 2(S1) + 1(S2) + 3(S3)
where:
S# = Taxa richness at each level of Beck's classifications (0-3).}
}
\value{
Beck's Biotic Index produces weighted richness values, which favors
organisms sensitive to degradation (Beck Classes 0 and 1).
}
\description{
Beck's Index
}
