\name{benthos-package}
\alias{benthos-package}
\alias{benthos}
\encoding{UTF-8}
\docType{package}
\title{
ecosystem functions, traits, and compositions of (mainly) aquatic benthic invertebrates 
}
\description{
Provides functions to lookup species 
 traits and compositions of (mainly) aquatic benthic invertebrates 
 and to calculate ecosystem functions of theese utilising Toms arificial neural network models 

}
\details{
\tabular{ll}{
Package: \tab benthos\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-15\cr
License: \tab GPL (>= 2)\cr
}
Usually, the methods are fed with species names or names of higher taxonomic units. Sometimes, more parameters are needed, e. g., mass, temperature, and waterdepth for the respiration and productivity neural network models.
}
\author{
Jan M Holstein <holstein@uni-bremen.de>

}
\references{
T. Brey, C. Müller-Wiegmann, Z. Zittier, W. Hagen, 2010. Body composition in aquatic organisms--a global data bank of relationships between mass, element composition 

Brey T, 2001. Population dynamics in benthic invertebrates. http://www.thomas-brey.de/science/
}
\keyword{ package }
