% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{benthos_cheat}
\alias{benthos_cheat}
\title{Benthos Cheat}
\usage{
benthos_cheat(long.df)
}
\arguments{
\item{long.df}{Taxonomic counts in a long data format.}
}
\value{
The Benthos package requires eight columns to exist before the metrics
can be calculated.  Although it is generally not recommended, five of these
columns (i.e., "STATION_ID", "AGENCY_CODE", "DATE", "METHOD", and
"SAMPLE_NUMBER") can be automatically generated and filled with a character
 sting (i.e., "blank"). Some of the required columns are not applicable to 
 all users and would have to be filled with a constant before proceeding;
 this function does the job for you.
}
\description{
Benthos Cheat
}
