% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metrics.R
\name{blank_col}
\alias{blank_col}
\title{Temporary vector with zero values}
\usage{
blank_col(taxon, taxa.rank)
}
\arguments{
\item{taxon}{The name of the taxon necessary for calculating a metric but
is not found in the data set.}

\item{taxa.rank}{Taxanomic taxa.rank (e.g. Class, Order, Family, Genus, etc.).}
}
\value{
Creates a temporary vector containing zero values the length of the
data frame specified by the object taxa.rank.  Prevents errors created by missing taxon names
necessary for metric calculations. For example, if no Trichoptera taxa were observed
in the data set then there would be no column named "Trichoptera" in the wide order
taxa.rank data frame.  When metrics, such as pct_ept or pct_trichoptera, search for a column
named "Trichoptera" a null value is returned and the function fails. Therefore,
this function temporarly fills the missing columns with zeros and the metric
can be calculated.
}
\description{
Temporary vector with zero values
}
