% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{bray_curtis}
\alias{bray_curtis}
\title{Bray-Curtis Dissimilarity}
\usage{
bray_curtis(n1, n2)
}
\arguments{
\item{n1}{abundances of species at site 1}

\item{n2}{abundances of species at site 2}
}
\value{
Bray-Curtis dissimilarity (0..1, 0 = equal, 1 = different)
}
\description{
Bray-Curtis Dissimilarity
}
\note{
species in n1 and n2 need to be aligned
}
\examples{
     n1 <- c(11,  0, 7,  8, 0)
     n2 <- c(24, 37, 5, 18, 1)
     bray_curtis(n1, n2)
}
