% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_taxa_metrics.R
\name{calc_pct_taxa}
\alias{calc_pct_taxa}
\title{The percentage each taxon makes up of a sample}
\usage{
calc_pct_taxa(wide.df, taxa.rank, master.df = BIBI::master)
}
\arguments{
\item{wide.df}{= Data in a wide data format.}

\item{taxa.rank}{= Taxonomic rank (all caps).}

\item{master.df}{= Master taxa list}
}
\value{
The percent of the sample represented by each taxa per taxon level.
}
\description{
The percentage each taxon makes up of a sample
}
