% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{calc_richness}
\alias{calc_richness}
\title{Internal Use: Calculate Richness}
\usage{
calc_richness(wide.df)
}
\arguments{
\item{wide.df}{a wide dataframe of taxonomic counts.}
}
\value{
Calculates richness based on the appropriate number of columns.
}
\description{
Internal Use: Calculate Richness
}
