% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{clean_up}
\alias{clean_up}
\title{Data Clean Up}
\usage{
clean_up(x)
}
\arguments{
\item{x}{= a data frame.}
}
\value{
capitalizes all of the column names, as well as, removes any leading
or trailing white space from the column names, columns of class character, 
and columns of class factor.
}
\description{
Data Clean Up
}
