\name{conv}
\alias{conv}
\encoding{UTF-8}
\docType{data}
\title{
conversion database
}
\description{
	The \emph{conversion database} consists of taxonomic units, 
	mostly species of mostly of benthic invertebrates,  
	taxonomic information and conversion factors as well as the respective short reference. 
}
\usage{data(conv)}
\format{
  A data frame with 5768 observations on the following 36 variables. 
}

\examples{
data(conv)

}
\keyword{datasets}
