% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Data Prep}
\usage{
data_prep(long.df, master.df)
}
\arguments{
\item{long.df}{Taxonomic counts in a long data format.}

\item{master.df}{A Master Taxa List containing the appropriate and applicable
taxonomic hierarchy and taxonomic traits. Specify 'master' to use the Master
Taxa List built in to the Benthos package.}
}
\value{
The provided taxonomic counts are merged with the provided Master 
Taxa List taxonomic hierarchy.
}
\description{
Data Prep
}
