% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drbc_data_prep.R
\name{drbc_data_prep}
\alias{drbc_data_prep}
\title{Prepare the DRBC Data}
\usage{
drbc_data_prep(taxa.df)
}
\arguments{
\item{taxa.df}{= Taxonomic counts in a long data format.}
}
\value{
Prepare taxonomic data from the Delaware River Basin Commission 
(DRBC).
}
\description{
Prepare the DRBC Data
}
