% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drbc_data_prep.R
\name{drbc_master_prep}
\alias{drbc_master_prep}
\title{Prepare the DRBC Attributes Table}
\usage{
drbc_master_prep(taxa.df)
}
\arguments{
\item{taxa.df}{= Taxonomic counts in a long data format.}
}
\value{
Prepare taxa attributes table from the data provided by the Delaware 
River Basin Commission (DRBC).
}
\description{
Prepare the DRBC Attributes Table
}
