% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{effective_richness}
\alias{effective_richness}
\title{Effective Number of Taxa}
\usage{
effective_richness(taxa.wide, index)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}

\item{index}{= Requires "shannon" (Shannon Wiener Diversity Index) or
"invsimpson" (Inverse Simpson's Diversity Index).}
}
\value{
This functions converts the common diversity indices values to
effective richness, as outlined by Jost (2006).  Effective richness is a
richness value weighted by species evenness; the value represents the taxa
richness of a perfectly even assemblage reuquired to produce an equivalent
diveristy index value to the observed assemblage. Shannon Wiener Diversity
is converted using the exponential function: exp(H')
where:
H'= Shannon Wiener Diversity

Inverse Simpsons Diveristy is converted using:
1/(1-(1/D))
where:
D = vegan Inverse Simpson's Diversity Index  (1/(sum(n/N)^2) )
n = the abundance of a particular taxon
N = the total abundance of organisms
}
\description{
Effective Number of Taxa
}
