% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{ept_rich_no_tol}
\alias{ept_rich_no_tol}
\title{EPT Richness Excluding Tolerant Taxa}
\usage{
ept_rich_no_tol(long, rank = "FAMILY", master, tolerance_value = "BIBI_TV")
}
\arguments{
\item{long}{Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{The taxonomic rank used to perform the analysis. You must
sepecify either 'FAMILY' or "GENUS.'}

\item{master}{A master taxa list including taxonomic ranks Phylum through
the specified taxonomic rank (Family or Genus) and the an
associated list of tolerance values. The default is set to the master taxa
list included in the BIBI package.  The master taxa list can be viewed with
the following script: master.df <- data(master)}

\item{tolerance_value}{The name of the column in the master taxon list
(specified using the master variable) that contains tolerance values on
a scale of 0-10.  Tolerant organisms are classified as organisms with a
tolerance value >= 7. The defualt is set to the the BIBI tolerance values,
which are tolerance values summarized from multiple sources.}
}
\value{
The number of taxa identified as EPT (Orders: Ephemeroptera,
Plecoptera, and Trichoptera), excluding taxa with a tolerance value >= 7.
}
\description{
EPT Richness Excluding Tolerant Taxa
}
