% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairfax_data_prep.R
\name{fairfax_attributes_prep}
\alias{fairfax_attributes_prep}
\title{Prepare the Fairfax County Taxonomic Attributes Table}
\usage{
fairfax_attributes_prep(attributes.df)
}
\arguments{
\item{attributes.df}{Taxonomic attributes table.}
}
\value{
Prepare Fairfax County's Taxonomic attribute table for use in the
Benthos package.
}
\description{
Prepare the Fairfax County Taxonomic Attributes Table
}
