% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairfax_data_prep.R
\name{fairfax_data_prep}
\alias{fairfax_data_prep}
\title{Prepare the Fairfax County Data}
\usage{
fairfax_data_prep(taxa.df)
}
\arguments{
\item{taxa.df}{Taxonomic counts in a long data format.}
}
\value{
Prepare taxonomic data from the Fairfax County, VA.
}
\description{
Prepare the Fairfax County Data
}
