% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{fill_taxa}
\alias{fill_taxa}
\title{Fill empty rows with lowest level of}
\usage{
fill_taxa(Taxon_List)
}
\arguments{
\item{Taxon_List}{= Master taxon list in wide format}
}
\value{
Fills in blank spaces with the lowest level of taxonomic
identification.
}
\description{
Fill empty rows with lowest level of
}
