% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_ambi}
\alias{get_ambi}
\title{Get Supplementary AMBI Sensitivity Groups}
\usage{
get_ambi(which = "NL")
}
\arguments{
\item{which}{which AMBI supplement? Currently only the Dutch supplement is
available (\code{which} = "NL")}
}
\value{
a data frame with columns \code{TAXON} containing taxa and 
    \code{GROUP} containing Dutch AMBI-groups
}
\description{
This function gets sensitivity groups that are supplementary to the AMBI of
Borja et al., (2000)
}
\references{
Borja, A., J. Franco and V. Perez, 2000. A Marine Biotic Index 
to Establish the Ecological Quality of Soft-Bottom Benthos Within 
European Estuarine and Coastal Environments. 
Marine Pollution Bulletin 40:1100-1114
}
