% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_iti}
\alias{get_iti}
\title{Get Infaunal Trophic Index}
\usage{
get_iti()
}
\value{
a data frame with columns \code{TAXON} containing taxa and 
    \code{GROUP} containing the ITI-groups of Gittenberger & Van Loon (2013).
}
\description{
This function gets the sensitivity groups to estimate the infaunal 
trophic index of Gittenberger et al., (2011)
}
\references{
Gittenberger A. and  W. van Loon, 2013. 
    Sensitivities of marine macrozoobenthos to environmental pressures 
    in the Netherlands. Nederlandse Faunistische 
    Mededelingen 41: 79-112.
}
