% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{gold}
\alias{gold}
\title{Proportion of Gastropoda, Oligochaeta, and Dipteran Individuals}
\usage{
gold(class.wide, order.wide)
}
\arguments{
\item{class.wide}{Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{order.wide}{Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
One minus the proportional number of gastropods (Class: Gastropoda),
oligochaetes (Class: Oligochaeta), and dipteran (Order: Diptera) individuals.
This metric typically decreases with degradation.
}
\description{
Proportion of Gastropoda, Oligochaeta, and Dipteran Individuals
}
