% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metrics.R
\name{group_taxa}
\alias{group_taxa}
\title{Data Frame of a specific list of taxa}
\usage{
group_taxa(NameList, Taxa.df)
}
\arguments{
\item{NameList}{= uninque list of taxa.}

\item{Taxa.df}{= Wide data frame format of taxonomic counts.}
}
\value{
Data frame of a specific list of taxa. Used in habit and functional
feeding group (FFG) fuctions to extract only the taxa specified in the object
NameList from the wide data frame of taxa (i.e. Family taxa.rank or Genus taxa.rank)
}
\description{
Data Frame of a specific list of taxa
}
