% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{hurlberts_pie}
\alias{hurlberts_pie}
\title{Probability of Interspecific Encounter (PIE)}
\usage{
hurlberts_pie(taxa.wide)
}
\arguments{
\item{taxa.wide}{Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}
}
\value{
Hurlbert's (1971) Probability of Interspecific Encounter (PIE):
PIE = (N/(N-1)) * (1 - sum(p^2))
where:
p = the propotion a specific taxon composes of a sample
N = the total abundance of organisms

This measurement is equivalent to Simpson's Index but includes a correction
factor base on the total abundance of organisms.
}
\description{
Probability of Interspecific Encounter (PIE)
}
