% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_azoic}
\alias{is_azoic}
\title{Test for Azoic Samples}
\usage{
is_azoic(x)
}
\arguments{
\item{x}{character vector containing taxa}
}
\value{
logical vector, with elements \code{TRUE} for azoic samples, 
     and \code{FALSE} otherwise.
}
\description{
Case-insensitive test for taxa starting with 'azoi'
}
\details{
Azoic samples need special attention during data analysis. 
     They should be marked as 'azoic', and taken care of during
     analysis. Note that an azoic sample is not the same as a
     record where a taxon has zero counts. The latter should be
     removed from further analysis, whereas the former provides
     important information.
}
