\name{itis}
\alias{itis}
\encoding{UTF-8}
\docType{data}
\title{
itis database
}
\description{
The \emph{itis database} is derived from ITIS (see source) and consists of an excerpt of taxonomic tree information for the North Sea benthic invertebrates.
}
\usage{data(itis)}
\format{
  A data frame with 1262 observations on the following 19 variables. 
}
\source{
\url{http://www.itis.gov/}
}

\examples{
data(itis)
## maybe str(itis) ; plot(itis.db) ...
}
\keyword{datasets}
