% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{long}
\alias{long}
\title{Long Data Frame
Transform taxa count data from wide to a long format.}
\usage{
long(wide.df, taxa.rank = "FAMILY")
}
\arguments{
\item{x}{Taxa count data}

\item{Level}{Taxonomic Level (PHYLUM, CLASS, ORDER, FAMILY, GENUS)}
}
\value{
Taxa counts in a long data format.
}
\description{
Long Data Frame
Transform taxa count data from wide to a long format.
}
