% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{menhinicks}
\alias{menhinicks}
\title{Menhinick's Index}
\usage{
menhinicks(taxa.wide)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}
}
\value{
Menhinick's Index: S/sqrt(N)
where:
S = taxa richness
N = the total number of individuals observed

Taxa richness is calculated using the vegan package specnumber function.
}
\description{
Menhinick's Index
}
