% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_northsea.R
\docType{data}
\name{northsea}
\alias{northsea}
\title{MWTL North Sea Bentos Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24983 rows and 9 columns.
}
\usage{
northsea
}
\description{
MWTL North Sea Bentos Data
}
\keyword{datasets}
