% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nysdec_data_prep.R
\name{nysdec_data_prep}
\alias{nysdec_data_prep}
\title{Prepare the NYSDEC SBU Data}
\usage{
nysdec_data_prep(Taxa.df)
}
\arguments{
\item{Taxa.df}{= A data frame containing raw taxonomic counts.}
}
\value{
Prepare taxonomic data from the New York State Department of 
Environmental Conservation (NYSDEC) Stream Biomonitoring Unit (SBU).
}
\description{
Prepare the NYSDEC SBU Data
}
