% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nysdec_master_prep.R
\name{nysdec_master_prep}
\alias{nysdec_master_prep}
\title{Prepare the NYSDEC Master Taxa List}
\usage{
nysdec_master_prep(dec.master)
}
\arguments{
\item{dec.master}{= A Master Taxa List containing the appropriate and applicable
taxonomic hierarchy and taxonomic traits. Specify 'master' to use the Master
Taxa List built in to the Benthos package.}
}
\value{
Prepare master taxa list from the New York State Department of 
Environmental Conservation (NYSDEC) Stream Biomonitoring Unit (SBU).
}
\description{
Prepare the NYSDEC Master Taxa List
}
