% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_metrics.R
\name{old_all_metrics}
\alias{old_all_metrics}
\title{Run All Metrics}
\usage{
old_all_metrics(master, long.df, taxa.rank = "FAMILY", rare = "NON_RARE",
  sample.size = 100, pct_un = NULL, bibi.standard = FALSE, seed = FALSE)
}
\arguments{
\item{master}{= The master taxa list contains taxonomic ranks from Phylum
to species and known taxonomic attributes.}

\item{long.df}{= Taxonomic counts in a long data format.}

\item{taxa.rank}{= The lowest taxonomic rank ("ORDER", "FAMILY", or "GENUS")
used to calculate the metrics.  If the majority of your taxa are identified
to the family level, then it would be inappropriate to perform metric
calculations at the genus level.}

\item{rare}{= Sample counts can be rarefied or unaltered. The default,
"NON_RARE," does not alter sample counts. "ALL_RARE" and "DIV_RARE" will
rarefy sample counts to the pecified sample size (sample.size) using
expected rarefaction. "ALL_RARE" use the rarefied counts to calulate all
of the metrics. "DIV_RARE" will only use the rarefied sample counts to
calculate richness and diversity metrics, all other metrics will be calculated
with the original counts.  "VEG_RARE" will rarefy the data using the vegan
package function rrarefy.  rrarefy produces a random rarefied sample. I
argue that we can use hypergeometric formulas to provide the expected number
of each taxon that will be selected at specified sample size.  "ALL_RARE" and
"DIV_RARE" may produce sample counts abouve and below the specifed sample size.
This is a result of rounding the taxonomic counts to the nearest integer. The
expected rarefied value is rarely produces an integer and it is inappropriate
to report a portion of an individual (e.g., 1.52 individuals).  Therefore,
the final expected rarefied value is rounded to the nearest integer.}

\item{sample.size}{= The rarefaction subsample size.  The sample size should be
smaller than the majority of the sample counts.  If a sample count is less than
or equal to the specified sample size, the sample will not be altered.}

\item{bibi.standard}{= Indicate TRUE if you would like to standardize your
taxa following the practices used in the Chessie BIBI.  The default, FALSE,
will not attempt to standardize the taxa.}

\item{seed}{= if true set the seed for reproducible results.}
}
\value{
Calculates all of applicable and available metrics in the package.
}
\description{
Run All Metrics
}
