% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_metrics.R
\name{pct_acid_tol}
\alias{pct_acid_tol}
\title{Acid Tolerance Index (ATI)}
\usage{
pct_acid_tol(genus.wide)
}
\arguments{
\item{genus.wide}{Taxonomic counts aggregated at the genus level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The number of Acid Tolerant Individuals (ATI) as described
by the NYSDEC (2014). The following genera are considered ATI taxa:
Epeorus (Ephemeroptera; Heptageniidae), Amphinemura (Plecoptera; Nemouridae),
 Leuctra (Plecoptera; Leuctridae), Isoperla (Plecoptera; Perlodidae),
  Rhyacophila (Trichoptera; Rhyacophilidae), Simulium (Diptera; Simuliidae),
Conchapelopia (Diptera; Chironomidae), Cricotopus (Diptera; Chironomidae),
Eukiefferiella (Diptera; Chironomidae), and Heterotrissocladius (Diptera; Chironomidae).
}
\description{
Acid Tolerance Index (ATI)
}
