% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_amphipoda}
\alias{pct_amphipoda}
\title{Percentage of Amphipod Individuals}
\usage{
pct_amphipoda(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as amphipods
(Order: Amphipoda).  This metric will typically increase with degradation.
}
\description{
Percentage of Amphipod Individuals
}
