% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_annelid_chiro}
\alias{pct_annelid_chiro}
\title{Percentage of Annelid and Chironomid Individuals}
\usage{
pct_annelid_chiro(phylum.wide, family.wide)
}
\arguments{
\item{phylum.wide}{Taxonomic counts aggregated at the phylum level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{family.wide}{Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as annelids (Phylum: Annelida)
and chironomids (Family: Chironomidae).  This metric typically increases with
degradation.
}
\description{
Percentage of Annelid and Chironomid Individuals
}
