% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metrics.R
\name{pct_attribute}
\alias{pct_attribute}
\title{The percent of a group}
\usage{
pct_attribute(Taxa.df, master.df, Group, Group_taxa.rank,
  taxa.rank = "FAMILY")
}
\arguments{
\item{Taxa.df}{= Wide data frame format of taxonomic counts.}

\item{Group}{= The taxonomic group to be assessed}

\item{Group_taxa.rank}{= The specific taxa.rank or taxa.ranks of the group to be assessed.}

\item{taxa.rank}{= The taxonomic taxa.rank used during the assessment.}

\item{master.df.df}{= taxonomic attributes table.}
}
\value{
The percentage of taxa representing a predefined group. Typically,
this function is used to assess functional feeding group and habits.
}
\description{
The percent of a group
}
