% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_bivalvia}
\alias{pct_bivalvia}
\title{Percentage of Bivalve Individuals}
\usage{
pct_bivalvia(class.wide)
}
\arguments{
\item{class.wide}{= Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as bivalves
(Class: Bivalvia).
}
\description{
Percentage of Bivalve Individuals
}
