% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_metrics.R
\name{pct_cc_chironomidae}
\alias{pct_cc_chironomidae}
\title{Percentage of Chironomids Identified as Chironomus and Cricoptopus}
\usage{
pct_cc_chironomidae(family.wide, genus.wide)
}
\arguments{
\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{genus.wide}{= Taxonomic counts aggregated at the genus level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of chironomids (Family: Chironomidae) individuals
that were identified as Chironomus (Diptera; Chironomidae) and
Cricotopus (Diptera; Chironomidae). This metric was found in the
Macroinvertebrate Metric Calculation Guidelines provided by GADNR/EPD
Watershed Protection Branch.
}
\description{
Percentage of Chironomids Identified as Chironomus and Cricoptopus
}
