% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_chironomidae}
\alias{pct_chironomidae}
\title{Percentage of Chironomid Individuals}
\usage{
pct_chironomidae(family.wide)
}
\arguments{
\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as chironomids
(Family: Chironomidae).
This metric will typically increase with degradation.
}
\description{
Percentage of Chironomid Individuals
}
