% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_coleoptera}
\alias{pct_coleoptera}
\title{Percentage of Coleopteran Individuals}
\usage{
pct_coleoptera(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as coleopterans
(Order: Coleoptera).
}
\description{
Percentage of Coleopteran Individuals
}
