% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_corbiculidae}
\alias{pct_corbiculidae}
\title{Percentage of Corbiculid Individuals}
\usage{
pct_corbiculidae(family.wide)
}
\arguments{
\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as corbiculids
(Family: Corbiculidae).
}
\description{
Percentage of Corbiculid Individuals
}
