% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_cote}
\alias{pct_cote}
\title{Percentage of COTE Individuals}
\usage{
pct_cote(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as coleopterans
(Order: Coleoptera), odonates (Order: Odonata), trichopterans
 (Order: Trichoptera), and ephemeropterans (Order: Ephemeroptera) (COTE).
 These organisms were used by Kamman (2007) as a lentic benthic
 macroinverterbate metric.

 Citation:
 KAMMAN, N. 2007. Development of Biocriteria for Vermont and
  New Hampshire Lakes Criteria Development for Macroinvertebrates
  for Three Lake Classes and Implementation Procedure for Biological
  Assessment of Vermont Lakes. Pages 1–21. Vermont Department
  of Environemental Conservation, Waterbury, VT.
}
\description{
Percentage of COTE Individuals
}
