% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_diptera}
\alias{pct_diptera}
\title{Percentage of Dipteran Individuals}
\usage{
pct_diptera(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as dipterans
(Order: Diptera). This metric will typically increase with degradation.
}
\description{
Percentage of Dipteran Individuals
}
