% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{pct_dom}
\alias{pct_dom}
\title{Percentage of the Most Dominant Taxon (Taxa)}
\usage{
pct_dom(taxa.wide, dom.level)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}

\item{dom.level}{= A numeric value 1-5 indicating the number}
}
\value{
Percent of individuals that represent the most abundant taxon or taxa.
dom.level can be used to specify 1st-5th most abundant taxa by specifying
the corresponding numeric value (1-5).  Values >1 include all of the previous
dominance levels. For example, dom.level = 3 is the percentage of the most
dominant taxon, the second most dominant taxon, and the third most
dominant taxon. This measure is related to taxa evenness. Typically
degradation is associated with elevated levels of the
most dominant taxon (taxa); therefore, this metric typically increases
with degradation.
}
\description{
Percentage of the Most Dominant Taxon (Taxa)
}
