% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metrics.R
\name{pct_dom1_group}
\alias{pct_dom1_group}
\title{The percent of the most dominant group}
\usage{
pct_dom1_group(long.df, master.df, Group, taxa.rank)
}
\arguments{
\item{long.df}{= long.df data frame format of taxonomic counts.}

\item{master.df}{= taxonomic attributes table.}

\item{Group}{= The taxonomic group to be assessed}

\item{taxa.rank}{= The taxonomic taxa.rank used during the assessment.}
}
\value{
The percentage of taxa representing by the most dominant (abundant)
group. Typically, this function is used to assess the functional feeding
groups and habits.
}
\description{
The percent of the most dominant group
}
