% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_ephemeroptera}
\alias{pct_ephemeroptera}
\title{Percentage of Ephemeropteran Individuals}
\usage{
pct_ephemeroptera(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as Ephemeropterans
(Order: Ephemeroptera).  This metric typically decreases with degradation.
}
\description{
Percentage of Ephemeropteran Individuals
}
