% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_ept}
\alias{pct_ept}
\title{Percentage of EPT Individuals}
\usage{
pct_ept(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as ephemeropterans
(Order: Ephemeroptera), plecopterans (Order: Plecoptera),
and trichopterans (Order: Trichoptera) (EPT).  This metric will typically
decrease with degradation.
}
\description{
Percentage of EPT Individuals
}
