% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_metrics.R
\name{pct_ept_hydropsyche}
\alias{pct_ept_hydropsyche}
\title{Percent EPT minus Hydropsyche}
\usage{
pct_ept_hydropsyche(order.wide, genus.wide)
}
\arguments{
\item{order.wide}{Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{genus.wide}{Taxonomic counts aggregated at the genus level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as EPT (Orders: Ephemeroptera,
Plecoptera, and Trichoptera) minus Hydropsyche (Trichoptera; Hydropsychidae)
individuals.
}
\description{
Percent EPT minus Hydropsyche
}
