% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{pct_ept_rich}
\alias{pct_ept_rich}
\title{EPT Percent Taxa Richness}
\usage{
pct_ept_rich(long, rank)
}
\arguments{
\item{long}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{= The taxonomic rank used to perform the analysis. You must
sepecify either 'FAMILY' or "GENUS.'}
}
\value{
The percentage of taxa identified as EPT (Orders: Ephemeroptera,
Plecoptera, and Trichoptera).
}
\description{
EPT Percent Taxa Richness
}
