% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_hydro_ept}
\alias{pct_hydro_ept}
\title{Percentage EPT Individuals Identified as Hydropsychids}
\usage{
pct_hydro_ept(order.wide, family.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of EPT (Orders: Ephemeroptera, Plecoptera,
and Trichoptera) individuals identified as hydropsychids
(Family: Hydropsychidae).  This metric typically increases with degradation.
}
\description{
Percentage EPT Individuals Identified as Hydropsychids
}
