% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_hydro_trichoptera}
\alias{pct_hydro_trichoptera}
\title{Percentage of Trichopterans Represented by Hydropsychids}
\usage{
pct_hydro_trichoptera(order.wide, family.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of trichopteran (Order:Trichoptera) individuals
that are hydropsychids (Family: Hydropsychidae). This metric typically
increases with degradation.
}
\description{
Percentage of Trichopterans Represented by Hydropsychids
}
