% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_hydropsychidae}
\alias{pct_hydropsychidae}
\title{Percentage of Hydropsychid Individuals}
\usage{
pct_hydropsychidae(family.wide)
}
\arguments{
\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as hydropsychids
(Family: Hydropsychidae).
}
\description{
Percentage of Hydropsychid Individuals
}
