% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_limestone}
\alias{pct_limestone}
\title{Percentage of Limestone Individuals}
\usage{
pct_limestone(order.wide, family.wide)
}
\arguments{
\item{order.wide}{order.wide = Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{family.wide}{Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as isopods (Order: Isopoda),
amphipods (Order: Amphipoda), and ephemerellids (Family: Ephemerellidae).
}
\description{
Percentage of Limestone Individuals
}
