% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_non_hydrop_trichoptera}
\alias{pct_non_hydrop_trichoptera}
\title{Percentage of Non-Hydropsychid Trichopteran Individuals}
\usage{
pct_non_hydrop_trichoptera(order.wide, family.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as trichopterans
(Order:Trichoptera) excluding hydropsychids (Family: Hydropsychidae).
}
\description{
Percentage of Non-Hydropsychid Trichopteran Individuals
}
