% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_non_insect}
\alias{pct_non_insect}
\title{Percentage of Non-Insect Individuals}
\usage{
pct_non_insect(class.wide)
}
\arguments{
\item{class.wide}{= Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals that were not identified as insects
(i.e., not identified as class Insecta).
}
\description{
Percentage of Non-Insect Individuals
}
