% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_metrics.R
\name{pct_non_tanytarsini_non_insecta}
\alias{pct_non_tanytarsini_non_insecta}
\title{Percentage of Non-Tanytarsini Diptera and Non-Insecta}
\usage{
pct_non_tanytarsini_non_insecta(class.wide, order.wide, tribe.wide)
}
\arguments{
\item{class.wide}{Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{order.wide}{Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{tribe.wide}{Taxonomic counts aggregated at the tribe level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of Diptera individuals not identified as Tanytarsini
(Diptera; Chironomidae) individuals plus the percentage of individuals not 
identified as Insecta. Tanytarsini is a tribe from the family 
Chironomidae. This metric was used by the Ohio Environmental Protection
Agency. Ohio Environmental Protection Agency. 1988. 
Biological criteria for the protection of aquatic life: Volume II: 
users manual for biological field assessment of Ohio surface waters. 
Ohio EPA, Division of Water Quality Monitoring and Assessment, 
Surface Water Section, Columbus.
}
\description{
Percentage of Non-Tanytarsini Diptera and Non-Insecta
}
