% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_odonata}
\alias{pct_odonata}
\title{Percentage of Odonate Individuals}
\usage{
pct_odonata(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as odonates (Order: Odonata).
}
\description{
Percentage of Odonate Individuals
}
