% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_oligo_chiro}
\alias{pct_oligo_chiro}
\title{Percentage of Oligochaet and Chironomid Individuals}
\usage{
pct_oligo_chiro(class.wide, family.wide)
}
\arguments{
\item{class.wide}{Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{family.wide}{Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as oligochaets
(Class: Oligochaeta) and chironomids (Family: Chironomidae).
}
\description{
Percentage of Oligochaet and Chironomid Individuals
}
