% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_oligochaeta}
\alias{pct_oligochaeta}
\title{Percentage of Oligochaet Individuals}
\usage{
pct_oligochaeta(class.wide)
}
\arguments{
\item{class.wide}{= Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as oligochaets
(Class: Oligochaeta).
}
\description{
Percentage of Oligochaet Individuals
}
