% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_metrics.R
\name{pct_orthocladiinae}
\alias{pct_orthocladiinae}
\title{Percentage of Orthocladiinae}
\usage{
pct_orthocladiinae(long)
}
\arguments{
\item{long}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}
}
\value{
The percentage of Orthocladiinae (Diptera; Chironomidae) individuals.
Orthocladiinae is a subfamily of the family Chironomidae.
}
\description{
Percentage of Orthocladiinae
}
