% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_plecoptera}
\alias{pct_plecoptera}
\title{Percentage of Plecopteran Individuals}
\usage{
pct_plecoptera(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as plecopterans
(Order: Plecoptera).
}
\description{
Percentage of Plecopteran Individuals
}
