% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_potec}
\alias{pct_potec}
\title{Percentage of POTEC Individuals}
\usage{
pct_potec(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as plecopterans
(Order: Plecoptera), odonates (Order: Odonata), trichopterans
 (Order: Trichoptera), ephemeropterans (Order: Ephemeroptera),
 and coleopterans (Order: Coleoptera) (POTEC).  This is an
 exploratory metric combining EPT taxa with COTE taxa.
 Coleopterans (Order: Coleoptera), odonates (Order: Odonata),
 trichopterans (Order: Trichoptera), and ephemeropterans
 (Order: Ephemeroptera) (COTE) were used by Kamman (2007) as
  a lentic benthic macroinverterbate metric.

 Citation:
 KAMMAN, N. 2007. Development of Biocriteria for Vermont and
  New Hampshire Lakes Criteria Development for Macroinvertebrates
  for Three Lake Classes and Implementation Procedure for Biological
  Assessment of Vermont Lakes. Pages 1–21. Vermont Department
  of Environemental Conservation, Waterbury, VT.
}
\description{
Percentage of POTEC Individuals
}
