% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_retreat_trichoptera}
\alias{pct_retreat_trichoptera}
\title{Percentage of Retreat-Making Trichopteran Individuals}
\usage{
pct_retreat_trichoptera(long)
}
\arguments{
\item{long}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}
}
\value{
The percentage of individuals identified as retreat-making
trichopterans (Suborder: Annulipalpia).
}
\description{
Percentage of Retreat-Making Trichopteran Individuals
}
