% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_simuliidae}
\alias{pct_simuliidae}
\title{Percentage of Simuliid Individuals}
\usage{
pct_simuliidae(family.wide)
}
\arguments{
\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as simuliids
(Family: Simuliidae).
}
\description{
Percentage of Simuliid Individuals
}
