% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_metrics.R
\name{pct_tanytarsini}
\alias{pct_tanytarsini}
\title{Percentage of Tanytarsini}
\usage{
pct_tanytarsini(tribe.wide)
}
\arguments{
\item{tribe.wide}{= Taxonomic counts aggregated at the tribe level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as Tanytarsini
(Diptera; Chironomidae) individuals. Tanytarsini is a tribe from the family
Chironomidae.
}
\description{
Percentage of Tanytarsini
}
