% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_taxon}
\alias{pct_taxon}
\title{Percentage of a Specified Taxon}
\usage{
pct_taxon(wide.df, taxon)
}
\arguments{
\item{wide.df}{Taxonomic counts aggregated at the appropriate taxonomic
rank in a wide data format. Use the wide function to prepare the data.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa 
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
A numeric vector containing the percentage of each sample 
represented by the specified taxon or taxa.
}
\description{
Calculate the percentage of each sample represented by the
specified taxon or taxa.
}
\details{
data format.
}
