% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance_metrics.R
\name{pct_tol_val}
\alias{pct_tol_val}
\title{Percentage of Intolerant Individuals}
\usage{
pct_tol_val(taxa.wide, master.df, tolerance.value = "BIBI_TV",
  lower.value = 0, upper.value = 3)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}

\item{master.df}{= A master taxa list including taxonomic ranks Phylum through
the specified taxonomic rank (Family or Genus) and the an
associated list of tolerance values. The default is set to the master taxa
list included in the BIBI package.  The master taxa list can be viewed with
the following script: master.df <- data(master)}

\item{tolerance.value}{= The name of the column in the master taxon list
(specified using the master variable) that contains tolerance values on
a scale of 0-10.}

\item{lower.value}{= the lower tolerance value bound.}

\item{upper.value}{= The upper tolerance value bound.}
}
\value{
The percentage of individuals that  fall within the specified
tolerance value bounds. Recommendations: Intolerant taxa
(lower.value = 0, upper.value = 3), facultative taxa
(lower.value = 4, upper.value = 6), and tolerant taxa
(lower.value = 7, upper.value = 10).
}
\description{
Percentage of Intolerant Individuals
}
