% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_trichoptera}
\alias{pct_trichoptera}
\title{Percentage of Trichopteran Individuals}
\usage{
pct_trichoptera(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as trichopterans
(Order: Trichoptera).
}
\description{
Percentage of Trichopteran Individuals
}
