% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_trichoptera_no_hydro}
\alias{pct_trichoptera_no_hydro}
\title{Percentage of Trichopterans Minus Hydropsychids}
\usage{
pct_trichoptera_no_hydro(order.wide, family.wide)
}
\arguments{
\item{order.wide}{Taxonomic counts aggregated at the class level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of trichopteran (Order:Trichoptera) individuals
minus the percentage of hydropsychids (Family: Hydropsychidae). 
This metric typically decreases with degradation.
}
\description{
Percentage of Trichopterans Minus Hydropsychids
}
