% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_unidentified}
\alias{pct_unidentified}
\title{Percentage of Unidentified Individuals}
\usage{
pct_unidentified(taxa.wide)
}
\arguments{
\item{taxa.wide}{Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}
}
\value{
The percentage of individuals that were not identified at the
specified taxonomic level.
}
\description{
Percentage of Unidentified Individuals
}
