% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{pct_unionoida}
\alias{pct_unionoida}
\title{Percentage of Unionoid Individuals}
\usage{
pct_unionoida(order.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The percentage of individuals identified as unionoids
(Order: Unionoida).
}
\description{
Percentage of Unionoid Individuals
}
