% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance_metrics.R
\name{pct_urban_intol}
\alias{pct_urban_intol}
\title{***Percentage of Urban Intolerant Individuals***}
\usage{
pct_urban_intol(long.df, master.df)
}
\arguments{
\item{long.df}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{master.df}{= A master taxa list including taxonomic ranks Phylum through
the specified taxonomic rank (Family or Genus) and the an
associated list of tolerance values. The default is set to the master taxa
list included in the BIBI package.  The master taxa list can be viewed with
the following script: master.df <- BIBI::master}
}
\value{
The percentage of indviduals that are classified as urban intolerant
by... (Where did this come from? Was in the 2011 BIBI.).  Only available
at the family level.
}
\description{
***Percentage of Urban Intolerant Individuals***
}
