% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{pielou}
\alias{pielou}
\title{Pielou's Evenness}
\usage{
pielou(taxa.wide)
}
\arguments{
\item{taxa.wide}{Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}
}
\value{
Pielou's Evenness Index: H'/ ln(S)
where:
H' = Shannon Weiner Diversity
S = taxa richness

If species richness is <= 1, this function will return 0.
Shannon Wiener diversity is calculated using the vegan package function
diversity with the index set to "shannon". Taxa richness is calculated
using the vegan package function specnumber.
}
\description{
Pielou's Evenness
}
