% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pool}
\alias{pool}
\alias{.pool}
\title{Pooling}
\usage{
pool(sample_id = 1:length(area), area, target_area, max_try = 100L)

.pool(sample_id = 1:length(area), area, target_area, max_try = 100L)
}
\arguments{
\item{sample_id}{sample identifier}

\item{area}{sampling area of \code{sample_id} (in the same units as 
\code{target_area})}

\item{target_area}{vector of length 2 containing the lower and upper bound 
of the pooled area (same units as \code{area})}

\item{max_try}{maximum number of unsuccessful pooling tries before the 
algorithm gives up.}
}
\value{
vector with idenitifiers (integers) indicating the pool to which 
    each sample belongs (NA for samples that could not be pooled)
}
\description{
This function randomly assigns samples to pools of approximately equal area
}
\section{Functions}{
\itemize{
\item \code{.pool}: internal function not supposed to be called directly.
}}

