% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_taxa}
\alias{read_taxa}
\alias{get_taxa}
\alias{validate_taxa}
\title{Read and Validate Taxa Data}
\usage{
read_taxa(filename)

get_taxa()

validate_taxa(.data)
}
\arguments{
\item{filename}{name of taxa file}

\item{.data}{table in taxa-format}
}
\description{
This function reads files in the taxa format.
}
\details{
Taxa files have the following format:
 \describe{
	  \item{group}{taxonomic group}
   \item{provided}{provided taxon name}
   \item{accepted}{accepted taxon name}
   \item{level}{taxonomic level}
	}
	Other columns are allowed, but silently ingored.
}
\section{Functions}{
\itemize{
\item \code{get_taxa()}: get default taxa list (TWN list extended with species Southern North Sea)

\item \code{validate_taxa()}: validator for taxa-format

}}
