% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_taxa_metrics.R
\name{rich_by_rank}
\alias{rich_by_rank}
\title{Richness of All Taxon}
\usage{
rich_by_rank(long.df, low.res.rank, high.res.rank, master.df)
}
\arguments{
\item{long.df}{Data in a long data format}

\item{low.res.rank}{Lower resolution taxonomic rank.}

\item{high.res.rank}{Higher resolution taxonomic rank.}

\item{master.df}{= Master Taxa List.}
}
\value{
Sequences through each taxon at level of the specified low.res.rank
and calculates richness based on the observations at the specified
high.res.rank.
}
\description{
Richness of All Taxon
}
