% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{rich_cote}
\alias{rich_cote}
\title{COTE Richness}
\usage{
rich_cote(long, rank = "FAMILY")
}
\arguments{
\item{long}{Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{The taxonomic rank used to perform the analysis. You must
sepecify either 'FAMILY' or "GENUS.'}
}
\value{
The number of taxa identified as COTE (Orders: Coleoptera, Odonata,
Trichoptera, and Ephemeroptera).
}
\description{
COTE Richness
}
