% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_metrics.R
\name{rich_ephem_epeorus}
\alias{rich_ephem_epeorus}
\title{Ephemeropteran Richness minus Epeorus}
\usage{
rich_ephem_epeorus(long, genus.wide)
}
\arguments{
\item{long}{Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{genus.wide}{Taxonomic counts aggregated at the genus level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The number of genera identified as ephemeropteran taxa
minus the genus Epeorus (Ephemeroptera; Heptageniidae),
as described by the NYSDEC.
}
\description{
Ephemeropteran Richness minus Epeorus
}
