% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{rich_ept}
\alias{rich_ept}
\title{EPT Richness}
\usage{
rich_ept(long, rank = "FAMILY")
}
\arguments{
\item{long}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{= The taxonomic rank used to perform the analysis. You must
sepecify either 'FAMILY' or "GENUS.'}
}
\value{
The number of taxa identified as EPT (Orders: Ephemeroptera,
Plecoptera, and Trichoptera), excluding taxa with a tolerance value >= 7.
}
\description{
EPT Richness
}
