% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{rich_nco}
\alias{rich_nco}
\title{Non-Chironomid and Oligochaet Taxa Richness}
\usage{
rich_nco(long, rank = "GENUS")
}
\arguments{
\item{long}{Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{The taxonomic rank used to perform the analysis. This
function requires a rank below the Order level taxonomic classification.}
}
\value{
The number of taxa that were not identified as oligochaets
(Class: Oligochaeta) and chironomids (Family: Chironomidae).
}
\description{
Non-Chironomid and Oligochaet Taxa Richness
}
