% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{rich_plecoptera}
\alias{rich_plecoptera}
\title{Plecopteran Richness}
\usage{
rich_plecoptera(long, rank = "FAMILY")
}
\arguments{
\item{long}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{= The taxonomic rank used to perform the analysis. This
function requires a rank below the Order level taxonomic classification.}
}
\value{
The number of taxa identified as plecopterans (Order: Plecoptera).
}
\description{
Plecopteran Richness
}
