% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metrics.R
\name{rich_trichoptera}
\alias{rich_trichoptera}
\title{Trichopteran Richness}
\usage{
rich_trichoptera(long, rank = "FAMILY")
}
\arguments{
\item{long}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{= The taxonomic rank used to perform the analysis. This
function requires a rank below the Order level taxonomic classification.}
}
\value{
The number of taxa identified as tichopterans (Order: Trichoptera).
}
\description{
Trichopteran Richness
}
