% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_taxa_metrics.R
\name{seq_pct_taxa}
\alias{seq_pct_taxa}
\title{The percentage each taxon makes up of a sample}
\usage{
seq_pct_taxa(long.df, master.df)
}
\arguments{
\item{long.df}{= Data in a long data format}

\item{master.df}{= Master taxa list}
}
\value{
The percent of the sample represented by each taxon per taxonomic level.
}
\description{
The percentage each taxon makes up of a sample
}
