% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_taxa_metrics.R
\name{seq_taxa_rich}
\alias{seq_taxa_rich}
\title{Sequence Through and Calculate the Richness of All Taxon}
\usage{
seq_taxa_rich(long.df, rank = "GENUS", master.df)
}
\arguments{
\item{long.df}{Data in a long data format}

\item{rank}{The taxonomic resolution of your dataset.}

\item{master.df}{Master Taxa List.}
}
\value{
Sequences through each taxonomic hierarchy and taxon, calculating the
richness of tax at the specified taxonomic rank ('''rank''').
}
\description{
Sequence Through and Calculate the Richness of All Taxon
}
