% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{shannon}
\alias{shannon}
\title{Shannon Wiener Diversity Index}
\usage{
shannon(taxa.wide)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}
}
\value{
Shannon Wiener Diversity Index: -sum(p * ln(p))
where:
p = the propotion a specific taxon composes of a sample

Shannon Wiener diversity is calculated using the vegan package function
diversity with the index set to "shannon".
}
\description{
Shannon Wiener Diversity Index
}
