% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{simpsons}
\alias{simpsons}
\title{Simpson's Diversity Index}
\usage{
simpsons(taxa.wide)
}
\arguments{
\item{taxa.wide}{= Taxonomic counts aggregated at the specific taxonomic
classification (e.g., Order, Family, or Genus) in a wide data format.
 Use the wide function to prepare the data.}
}
\value{
Simpson's Divieristy Index: 1 - sum(n/N)^2
where:
n = the abundance of a particular taxon
N = the total abundance of organisms

Simpson's diversity is calculated using the vegan package function
diversity with the index set to "simpson".
}
\description{
Simpson's Diversity Index
}
