% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{taxon_richness}
\alias{taxon_richness}
\title{Taxon Richness}
\usage{
taxon_richness(long, taxon, low.res.rank, high.res.rank)
}
\arguments{
\item{long}{Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{rank}{The taxonomic rank used to perform the analysis. This
function requires a rank below the Order level taxonomic classification.}
}
\value{
The number of taxa identified as ephemeropterans (Order: Ephemeroptera).
}
\description{
Taxon Richness
}
