% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance_metrics.R
\name{tol_pres_abs}
\alias{tol_pres_abs}
\title{Tolerance Index for Presence/Absence}
\usage{
tol_pres_abs(long.df, taxa.rank = "FAMILY", master.df,
  tolerance.value = "BIBI_TV")
}
\arguments{
\item{long.df}{= Taxonomic counts arrange in a long data format (i.e., each
row represents a unique sample and taxon).}

\item{taxa.rank}{= The taxonomic rank used to perform the analysis. You must
sepecify either "FAMILY" or "GENUS" (Defualt = "FAMILY").}

\item{master.df}{= A master taxa list including taxonomic ranks Phylum through
the specified taxonomic rank (Family or Genus) and the an
associated list of tolerance values. The default is set to the master taxa
list included in the BIBI package.  The master taxa list can be viewed with
the following script: master.df <- data(master)}

\item{tolerance.value}{= The name of the column in the master taxon list
(specified using the master variable) that contains tolerance values on
a scale of 0-10.}
}
\value{
The average tolerance score using only presence/absence data for
each unique sampling event. This metric is calculated at the family or
genus level. If taxon does not have an assigned tolerance value it will
not contribute to the final score. The each taxon is represented by the
taxon's tolerance value.  If the tolerance value is
missing it would be treated effectively as a zero, which would reduce
the average tolerance score; to avoid this issue, it is more effective to remove
the taxa without any tolerance values. However, if a large portion of
the sample is composed of a taxon or taxa without assigned tolerance
values the final value may be a poor representation of the sample.
Every effort should be made to assign each taxon a tolerance value.
If you have tolerance values that you would like to contribute and share
through the BIBI package please email Zachary M. Smith (zsmith@icprb.org).
}
\description{
Tolerance Index for Presence/Absence
}
