% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{total_abundance}
\alias{total_abundance}
\alias{total_abundance_}
\alias{lnn}
\alias{lnn_}
\title{Total Abundance}
\usage{
total_abundance(.data = NULL, count, na.rm = FALSE)

total_abundance_(.data = NULL, count, na.rm = FALSE)

lnn(.data = NULL, count, na.rm = FALSE)

lnn_(.data = NULL, count, na.rm = FALSE)
}
\arguments{
\item{.data}{data in a \code{data.frame}, \code{tibble}, 
\code{data.table}, database etc.}

\item{count}{counts (\code{numeric})}

\item{na.rm}{Should missing values (including \code{NaN}) 
be removed? (\code{logical})}
}
\value{
total number of individuals (\code{integer})
}
\description{
The total number of individuals.
}
\section{Functions}{
\itemize{
\item \code{total_abundance_()}: version suitable for calling from a function
(see package \pkg{lazyeval}).

\item \code{lnn()}: natural log of total abundance + 1
(see package \pkg{lazyeval}).

\item \code{lnn_()}: version of lnn suitable for calling from a function
(see package \pkg{lazyeval}).

}}
\examples{
 total_abundance(count = c(4, 6))
     
}
