% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{wide}
\alias{wide}
\title{Wide Data Frame
Transform taxa count data from long to wide format}
\usage{
wide(long.df, taxa.rank, pct.unid = NULL)
}
\arguments{
\item{taxa.rank}{= Taxonomic Level (PHYLUM, CLASS, ORDER, FAMILY, GENUS)}

\item{pct.unid}{= A threshold can be established to exclude samples that do
have too many unidentified taxa in a sample. Enter the percentage of the
sample that you are comfortable with being unidentified.  If not specified
the function will not excluded any samples.}

\item{long}{= Taxa count data}
}
\value{
Taxa counts in wide format
}
\description{
Wide Data Frame
Transform taxa count data from long to wide format
}
