% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{abundance}
\alias{abundance}
\alias{abundance_}
\title{Abundance}
\usage{
abundance(.data = NULL, taxon = NULL, count)

abundance_(.data = NULL, taxon = NULL, count)
}
\arguments{
\item{.data}{data in a \code{data.frame}, \code{data_frame}, 
\code{data.table}, database etc.}

\item{taxon}{name of column in \code{.data} containing taxa}

\item{count}{name of column in \code{.data} containing counts}
}
\value{
\code{numeric} vector with abundance per taxon.
}
\description{
The number of indiviuals in each taxon.
}
\section{Functions}{
\itemize{
\item \code{abundance_}: version suitable for calling from a function
(see package \pkg{lazyeval}).
}}

\note{
due to pooling, the abundance is not necessarily an integer
}
\examples{
 abundance(
     taxon = c("Euspira pulchella", "Nephtys cirrosa"), 
     count = c(4, 6)
 )
}
