% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_metrics.R
\name{all_metrics}
\alias{all_metrics}
\title{Run All Metrics}
\usage{
all_metrics(long.df, master.df, taxa.rank, tv.col = "BIBI_TV",
  ffg.col = "BIBI_FFG", hab.col = "BIBI_HABIT", beck.col = "BECK_CLASS")
}
\arguments{
\item{long.df}{Taxonomic counts in a long data format.}

\item{master.df}{The master taxa list contains taxonomic ranks from Phylum
to species and known taxonomic attributes.}

\item{taxa.rank}{The lowest taxonomic rank ("ORDER", "FAMILY", or "GENUS")
used to calculate the metrics.  If the majority of your taxa are identified
to the family level, then it would be inappropriate to perform metric
calculations at the genus level.}

\item{tv.cols}{The name of the tolerance value column you want to use
during metric calculations.}

\item{ffg.cols}{The name of the ffg value column you want to use
during metric calculations.}

\item{hab.cols}{The name of the habit value column you want to use
during metric calculations.}
}
\value{
Calculates all of applicable and available metrics in the package.
}
\description{
Run All Metrics
}
