% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{gold}
\alias{gold}
\title{Proportion of Gastropoda, Oligochaeta, and Dipteran Individuals}
\usage{
gold(long.df, unique.id.col, count.col, class.col, order.col)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{unique.id.col}{The name of the column that contains a unique sampling
event ID.}

\item{count.col}{The name of the column that contains taxanomic counts.}

\item{class.col}{The name of the column with class-level taxonomic information.}

\item{order.col}{The name of the column with order-level taxonomic information.}
}
\value{
A numeric vector containing the percentage of each sample 
represented by the specified taxon or taxa.

Numeric vector of percentages.
}
\description{
The percentage of individuals not represented by 
gastropods (Class: Gastropoda), oligochaetes (Class: Oligochaeta),
 and dipteran (Order: Diptera). This metric typically decreases with 
 degradation.
}
