% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nysdec_data_prep.R
\name{nysdec_data_prep}
\alias{nysdec_data_prep}
\title{Prepare the NYSDEC SBU Data}
\usage{
nysdec_data_prep(taxa.df, master.df)
}
\arguments{
\item{taxa.df}{Taxonomic counts in a long data format.}

\item{master.df}{A Master Taxa List containing the appropriate and applicable
taxonomic hierarchy and taxonomic traits. Specify 'master' to use the Master
Taxa List built in to the Benthos package.}
}
\value{
Prepare taxonomic data from the New York State Department of 
Environmental Conservation (NYSDEC) Stream Biomonitoring Unit (SBU).
}
\description{
Prepare the NYSDEC SBU Data
}
