% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_metrics.R
\name{pct_taxon}
\alias{pct_taxon}
\title{Percentage of Amphipod Individuals}
\usage{
pct_taxon(wide.df, taxon)
}
\arguments{
\item{wide.df}{= Taxonomic counts aggregated at the appropriate taxonomic
rank in a wide data format. Use the wide function to prepare the data.}

\item{taxon}{= The taxon of interest.}
}
\value{
The percentage of individuals identified as the specified taxon.
}
\description{
Percentage of Amphipod Individuals
}
